# Conflicts:
#	cmd/werf/common/conveyor_options.go
#	cmd/werf/lint/lint.go
#	docs/_includes/reference/cli/werf_build.md
#	docs/_includes/reference/cli/werf_compose_config.md
#	docs/_includes/reference/cli/werf_compose_down.md
#	docs/_includes/reference/cli/werf_compose_run.md
#	docs/_includes/reference/cli/werf_compose_up.md
#	docs/_includes/reference/cli/werf_export.md
#	docs/_includes/reference/cli/werf_kube_run.md
#	docs/_includes/reference/cli/werf_lint.md
#	docs/_includes/reference/cli/werf_run.md
#	go.mod
#	pkg/buildah/common.go
#	pkg/container_backend/buildah_backend.go
#	pkg/container_backend/docker_server_backend.go
#	pkg/container_backend/perf_check_container_backend.go
#	pkg/tmp_manager/gc.go
#	pkg/tmp_manager/gc_test.go
#	pkg/util/parallel/parallel.go
#	pkg/util/parallel/worker.go
#	test/pkg/werf/project.go
#	trdl_channels.yaml

+++ pkg/util/parallel/worker.go
@@ -4,8 +4,11 @@ import (
 	"fmt"
 	"os"
 	"sync/atomic"
+
+	"github.com/werf/werf/v2/pkg/tmp_manager"
 )
 
+// Worker
 // This is a worker for concurrent writing and reading logs.
 // It uses *os.File under the hood.
 //
@@ -13,24 +16,24 @@ import (
 // - the writer appends data only;
 // - the reader reads already appended data (or nothing).
 // Because of that, no race condition happens while accessing file data.
-type parallelWorker struct {
-	ID int
+type Worker struct {
+	readOffset  atomic.Int64
+	writeOffset atomic.Int64
+
+	halfClosed atomic.Bool
 
 	file *os.File
 
-	failed     atomic.Bool
-	halfClosed atomic.Bool
-
-	writeOffset atomic.Int64
-	readOffset  atomic.Int64
+	ID int
 }
 
 // Write implements io.Writer.
 // It appends to file and accumulates total write offset.
-func (w *parallelWorker) Write(p []byte) (int, error) {
+func (w *Worker) Write(p []byte) (int, error) {
 	if w.halfClosed.Load() {
-		return 0, fmt.Errorf("worker %d is half closed but tries to write: %s", w.ID, p)
+		return 0, fmt.Errorf("worker is half closed but tries to write: %s", p)
 	}
+
 	offset, err := w.file.Write(p)
 	w.writeOffset.Add(int64(offset))
 	return offset, err
@@ -39,14 +42,14 @@ func (w *parallelWorker) Write(p []byte) (int, error) {
 // Read implements io.Reader.
 // It reads a file and accumulates total read offset.
 // It resumes reading from "total read offset" and reads until EOF, where EOF is handled with os.File.
-func (w *parallelWorker) Read(p []byte) (int, error) {
+func (w *Worker) Read(p []byte) (int, error) {
 	offset, err := w.file.ReadAt(p, w.readOffset.Load())
 	w.readOffset.Add(int64(offset))
 	return offset, err
 }
 
 // HalfClose closes writing or returns error if already half-closed
-func (w *parallelWorker) HalfClose() error {
+func (w *Worker) HalfClose() error {
 	if w.halfClosed.Load() {
 		return fmt.Errorf("worker %d is already half closed", w.ID)
 	}
@@ -54,18 +57,8 @@ func (w *parallelWorker) HalfClose() error {
 	return nil
 }
 
-// Fail marks worker as failed
-func (w *parallelWorker) Fail() {
-	w.failed.Store(true)
-}
-
-// Failed returns true if worker failed
-func (w *parallelWorker) Failed() bool {
-	return w.failed.Load()
-}
-
 // Readable returns true if worker is readable
-func (w *parallelWorker) Readable() bool {
+func (w *Worker) Readable() bool {
 	if !w.halfClosed.Load() {
 		return true
 	}
@@ -74,7 +67,7 @@ func (w *parallelWorker) Readable() bool {
 
 // Close implements io.Closer closing tmp file.
 // It ensures that worker is half closed
-func (w *parallelWorker) Close() error {
+func (w *Worker) Close() error {
 	w.halfClosed.Store(true)
 
 	if err := w.file.Close(); err != nil {
@@ -84,20 +77,24 @@ func (w *parallelWorker) Close() error {
 }
 
 // Cleanup removes tmp file
-func (w *parallelWorker) Cleanup() error {
+func (w *Worker) Cleanup() error {
+	if !w.halfClosed.Load() {
+		return fmt.Errorf("worker %d is not half closed yet", w.ID)
+	}
+
 	if err := os.Remove(w.file.Name()); err != nil {
 		return fmt.Errorf("failed to remove tmp file for worker %d: %w", w.ID, err)
 	}
 	return nil
 }
 
-func newParallelWorker(id int) (*parallelWorker, error) {
-	file, err := os.CreateTemp("", fmt.Sprintf("parallel-worker-%d-%d-*.log", os.Getpid(), id))
+func NewWorker(id int) (*Worker, error) {
+	file, err := tmp_manager.TempFile(fmt.Sprintf("parallel-worker-%d-%d-*.log", os.Getpid(), id))
 	if err != nil {
 		return nil, fmt.Errorf("failed to create temp file for worker %d: %w", id, err)
 	}
 
-	return &parallelWorker{
+	return &Worker{
 		ID:   id,
 		file: file,
 	}, nil
